feat: added custom header animation logic
added custom header animation logic
