+++ src/pages/wallet/ui/wallet.tsx
@@ -1,32 +1,13 @@
 import React, { useCallback } from 'react'
-import {
-  FlatList,
-  LayoutChangeEvent,
-  ListRenderItem,
-  NativeScrollEvent,
-  NativeSyntheticEvent,
-} from 'react-native'
-import Animated, {
-  useAnimatedStyle,
-  useSharedValue,
-  interpolate,
-  Extrapolate,
-} from 'react-native-reanimated'
+import { FlatList, ListRenderItem } from 'react-native'
 
 import { Typography } from '@shared/ui/atoms'
 import { DefaultPageTemplate } from '@shared/ui/templates'
-import { useTheme, styled } from '@shared/ui/theme'
+import { useTheme } from '@shared/ui/theme'
 import { Header } from '@shared/ui/molecules'
+import { useHeaderAnimation } from '@shared/lib/hooks'
 
 import { mocks } from './mocks'
-
-const AnimatedTypography = Animated.createAnimatedComponent(Typography)
-
-const StyledAnimatedTypography = styled(AnimatedTypography)`
-  align-self: flex-start;
-  padding-top: ${({ theme }) => theme.spacing(1.25)}px;
-  padding-bottom: ${({ theme }) => theme.spacing(2.5)}px;
-`
 
 const keyExtractor = (item: string, index: number) => item + index
 
@@ -34,61 +15,11 @@ type TWalletProps = { header: string }
 
 export const Wallet = ({ header }: TWalletProps) => {
   const theme = useTheme()
-  const offsetY = useSharedValue(0)
-  const initialHeaderWidth = useSharedValue(0)
-
-  const typographyStyles = useAnimatedStyle(() => {
-    const scale = interpolate(
-      offsetY.value,
-      [0, -250],
-      [1, 1.2],
-      Extrapolate.CLAMP,
-    )
-
-    const headerWidth = initialHeaderWidth.value
-    const translateX = (headerWidth * scale - headerWidth) / 2
-
-    return {
-      transform: [
-        {
-          scale,
-        },
-        { translateX },
-      ],
-      opacity: interpolate(offsetY.value, [0, 10], [1, 0], Extrapolate.CLAMP),
-    }
-  })
-
-  const onLayout = useCallback(
-    (event: LayoutChangeEvent) => {
-      initialHeaderWidth.value = event.nativeEvent.layout.width
-    },
-    [initialHeaderWidth],
-  )
-
-  const onScroll = useCallback(
-    (event: NativeSyntheticEvent<NativeScrollEvent>) => {
-      offsetY.value = event.nativeEvent.contentOffset.y
-    },
-    [offsetY],
-  )
+  const { offsetY, onScroll, ListHeaderComponent } = useHeaderAnimation(header)
 
   const renderItem: ListRenderItem<string> = useCallback(
     ({ item }) => <Typography>{item}</Typography>,
     [],
-  )
-
-  const ListHeaderComponent = useCallback(
-    () => (
-      <StyledAnimatedTypography
-        onLayout={onLayout}
-        variant="h1"
-        style={typographyStyles}
-      >
-        {header}
-      </StyledAnimatedTypography>
-    ),
-    [header, onLayout, typographyStyles],
   )
 
   return (
