+++ src/shared/lib/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useHeaderAnimation } from './use-header-animation'
