+++ src/shared/lib/hooks/use-header-animation/index.ts
@@ -1,0 +1,1 @@
+export { useHeaderAnimation } from './use-header-animation'
